Fix ARM32 warnings
(cherry picked from commit e89b09e9eac4dfd4ac6673b9b7aaed9610466384)
(cherry picked from commit 6fc17804e8f5ee6769dd5a1dfb794852b66afe5c)

+++ mem/q_malloc.c
@@ -192,10 +192,10 @@ struct qm_block *qm_malloc_init(char *address, unsigned long size, char *name)
 
 	/* make address and size multiple of 8*/
 	start=(char*)ROUNDUP((unsigned long) address);
-	LM_DBG("QM_OPTIMIZE=%lu, /ROUNDTO=%lu, %zu-bytes aligned\n",
+	LM_DBG("QM_OPTIMIZE=%lu, /ROUNDTO=%lu, %lu-bytes aligned\n",
 			Q_MALLOC_OPTIMIZE, Q_MALLOC_OPTIMIZE/QM_ROUNDTO, QM_ROUNDTO);
-	LM_DBG("QM_HASH_SIZE=%lu, qm_block size=%lu, frag_size=%zu\n",
-			QM_HASH_SIZE, (long)sizeof(struct qm_block), FRAG_OVERHEAD);
+	LM_DBG("QM_HASH_SIZE=%lu, qm_block size=%zu, frag_size=%zu\n",
+			QM_HASH_SIZE, sizeof(struct qm_block), FRAG_OVERHEAD);
 	LM_DBG("params (%p, %lu), start=%p\n", address, size, start);
 	if (size<start-address) return 0;
 	size-=(start-address);
